ahora si que me vas a comer los huevos

+++ p4/pescalar_par2.c
@@ -10,8 +10,8 @@ int main(void)
 	float *A=NULL, *B=NULL;
 	long long k=0;
 	struct timeval fin,ini;
-	float sum=0;
-	
+	double sum=0;
+
 	A = generateVector(M);
 	B = generateVector(M);
 	if ( !A || !B )
@@ -21,7 +21,7 @@ int main(void)
 		freeVector(B);
 		return -1;
 	}
-	
+
 	gettimeofday(&ini,NULL);
 	/* Bloque de computo */
 	sum = 0;
@@ -33,7 +33,7 @@ int main(void)
 	/* Fin del computo */
 	gettimeofday(&fin,NULL);
 
-	printf("Resultado: %f\n",sum);
+	printf("Resultado: %lf\n",sum);
 	printf("Tiempo: %f\n", ((fin.tv_sec*1000000+fin.tv_usec)-(ini.tv_sec*1000000+ini.tv_usec))*1.0/1000000.0);
 	freeVector(A);
 	freeVector(B);
